test(infra/axiosadapter): refactor tests moving functions to your folders
