+++ tests/infra/helpers/extensionFactories.ts
@@ -1,0 +1,11 @@
+import { AxiosResponse } from 'axios';
+import { HttpResponse } from '~/data/http';
+
+export const toHttpResponse = (errorResponse: AxiosResponse) => {
+  const httpErrorResponse = {
+    statusCode: errorResponse?.status,
+    body: errorResponse?.data,
+  } as HttpResponse;
+
+  return httpErrorResponse;
+};
